Исправил обработку исключения.

+++ src/main/java/TASK_4/Delenie.java
@@ -3,10 +3,10 @@ package TASK_4;
  * @author Stolyarchuk Andrey
  */
 public class Delenie extends Operation {
-    public Delenie(double a, double b, int oper) {
+    public Delenie(int a, int b, int oper) {
         super(a, b, oper);
     }
-    public double doDel() {
+    public int doDel() {
         return getA() / getB();
     }
     @Override
